Pequeño cambio puntos

+++ asi-2425-notebook-rest-main/src/main/java/es/udc/asi/notebook_rest/model/domain/CampeonatoConstructor.java
@@ -11,7 +11,7 @@ public class CampeonatoConstructor {
   private Long id;
 
   @Column
-  private Long puntos;
+  private Float puntos;
   private Long victorias;
 
   @ManyToOne
@@ -29,7 +29,7 @@ public class CampeonatoConstructor {
 
   }
 
-  public CampeonatoConstructor(Long puntos, Long victorias) {
+  public CampeonatoConstructor(Float puntos, Long victorias) {
 
     this.puntos = puntos;
     this.victorias = victorias;
@@ -52,11 +52,11 @@ public class CampeonatoConstructor {
     this.victorias = victorias;
   }
 
-  public Long getPuntos() {
+  public Float getPuntos() {
     return puntos;
   }
 
-  public void setPuntos(Long puntos) {
+  public void setPuntos(Float puntos) {
     this.puntos = puntos;
   }
